lang/python/code/cmdbop/pyproject.toml
[project]
name = "cmdbop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "fastapi[standard]>=0.125.0",
    "orjson>=3.11.5",
    "pyyaml>=6.0.3",
    "sqlalchemy[asyncio]>=2.0.45",
]
